refactor(pos) :  패키지 - product 추가
