+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,15 +1,7 @@
 package kitchenpos.menu.domain;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.Transient;
+import jakarta.persistence.*;
+import kitchenpos.product.domain.Product;
 
 import java.util.UUID;
